Support for nuget package with multiple Roslyn version analyzers in CLI
