+++ src/NuGetForUnity/Editor/Helper/AnalyzerHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7abfc0bbece578541aa4b670879d6dc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
